rename: errorBoundry to errorBoundary
